# openapi.yaml
openapi: 3.1.0
info:
  title: Personal Fitness Agent API
  version: 1.0.0
  description: API for logging and retrieving fitness data (strength sets, workouts, readiness, etc.).
servers:
  - url: https://gpt-proxy-747xkkae2-dans-projects-19e6c8d0.vercel.app
paths:
  "/api/strength?x-vercel-protection-bypass=1CQBplG678yMDnEwGaTSEojJb2h1uUMm":
    post:
      summary: Log a strength-training set
      operationId: logStrengthSet
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StrengthSet'
      responses:
        "200":
          description: Successfully stored the set
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  stored:
                    $ref: '#/components/schemas/StrengthSet'
  "/api/trends?x-vercel-protection-bypass=1CQBplG678yMDnEwGaTSEojJb2h1uUMm":
    get:
      operationId: getWorkoutTrends
      summary: Get weekly workout trends
      security:
        - ApiKeyAuth: []
      parameters:
        - name: weeks
          in: query
          description: Number of recent ISO weeks to include (default 4)
          required: false
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Weekly trend summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  weeks:
                    type: array
                    items:
                      type: object
                      properties:
                        week: { type: string, example: "2025-W27" }
                        metrics:
                          type: object
                          additionalProperties:
                            type: object
                            properties:
                              tonnage: { type: number }
                              distance: { type: number }
                              duration: { type: number }
                              sessions: { type: integer }
                  progressing:
                    type: array
                    items: { type: string }
                  stalling:
                    type: array
                    items: { type: string }
  "/api/workout?x-vercel-protection-bypass=1CQBplG678yMDnEwGaTSEojJb2h1uUMm":
    post:
      summary: Log an endurance or cross-training workout (run, bike, swim, etc.)
      operationId: logWorkout
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Workout'
      responses:
        "200":
          description: Successfully stored the workout
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  stored:
                    $ref: '#/components/schemas/Workout'
  "/api/run-trends?x-vercel-protection-bypass=1CQBplG678yMDnEwGaTSEojJb2h1uUMm":
    get:
      summary: Get weekly running trends
      operationId: getRunTrends
      security:
        - ApiKeyAuth: []
      parameters:
        - name: weeks
          in: query
          description: Number of ISO weeks to include (default 4)
          required: false
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Weekly running trend summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  weeks:
                    type: array
                    items:
                      type: object
                      properties:
                        week: { type: string, example: "2025-W27" }
                        distance: { type: number, description: "Total km" }
                        duration: { type: number, description: "Total minutes" }
                        sessions: { type: integer }
                        avgPace: { type: number, nullable: true, description: "Minutes per km" }
                  progressing:
                    type: boolean
  "/api/cycle-trends?x-vercel-protection-bypass=1CQBplG678yMDnEwGaTSEojJb2h1uUMm":
    get:
      summary: Get weekly cycling trends
      operationId: getCycleTrends
      security:
        - ApiKeyAuth: []
      parameters:
        - name: weeks
          in: query
          description: Number of ISO weeks to include (default 4)
          required: false
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Weekly cycling trend summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  weeks:
                    type: array
                    items:
                      type: object
                      properties:
                        week: { type: string, example: "2025-W27" }
                        distance: { type: number, description: "Total km" }
                        duration: { type: number, description: "Total minutes" }
                        sessions: { type: integer }
                        avgSpeed: { type: number, nullable: true, description: "Km/h" }
                  progressing:
                    type: boolean
  "/api/swim-trends?x-vercel-protection-bypass=1CQBplG678yMDnEwGaTSEojJb2h1uUMm":
    get:
      summary: Get weekly swimming trends
      operationId: getSwimTrends
      security:
        - ApiKeyAuth: []
      parameters:
        - name: weeks
          in: query
          description: Number of ISO weeks to include (default 4)
          required: false
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Weekly swimming trend summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  weeks:
                    type: array
                    items:
                      type: object
                      properties:
                        week: { type: string, example: "2025-W27" }
                        distance: { type: number, description: "Total km" }
                        duration: { type: number, description: "Total minutes" }
                        sessions: { type: integer }
                        avgPace: { type: number, nullable: true, description: "Minutes per km" }
                  progressing:
                    type: boolean
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: |
        Fitness Agent API key. Also requires x-vercel-protection-bypass header.
        Example value: super-secret-123
  schemas:
    StrengthSet:
      type: object
      required: [date, exercise, setNumber, reps, weight]
      properties:
        date:
          type: string
          description: ISO date (YYYY-MM-DD)
          example: "2025-07-12"
        exercise:
          type: string
          example: Squat
        setNumber:
          type: integer
          example: 1
        reps:
          type: integer
          example: 8
        weight:
          type: number
          example: 135
        rpe:
          type: number
          nullable: true
          example: 7
        notes:
          type: string
          nullable: true
          example: "test via ChatGPT Action"
    Workout:
      type: object
      required: [date, type]
      properties:
        date:
          type: string
          description: ISO date (YYYY-MM-DD)
          example: "2025-07-12"
        type:
          type: string
          description: Activity type (Run, Bike, Swim, etc.)
          example: Run
        duration:
          type: number
          description: Minutes
          example: 45
        distance:
          type: number
          description: Kilometers
          example: 10
        calories:
          type: number
          example: 500
        sets:
          type: integer
          nullable: true
          example: 3
        weight:
          type: number
          nullable: true
          example: 0
        reps:
          type: integer
          nullable: true
          example: 0
        notes:
          type: string
          nullable: true
          example: "Afternoon tempo run" 